fix: logic auth

+++ src/resources/board/board.memory.repository.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-else-return */
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 /* eslint-disable @typescript-eslint/ban-types */
 import { getRepository } from 'typeorm';
@@ -35,18 +36,25 @@ const updateBoard = async (
 };
 
 const deleteBoard = async (id: string) => {
-  const tasksRepository = getRepository(Task);
-  const arrayOfTasks = await tasksRepository.find({ where: { boardId: id } });
-  if (arrayOfTasks.length > 0) {
-    for (let i = 0; i < arrayOfTasks.length; i += 1) {
-      const idTasks = arrayOfTasks[i]!.id;
-      tasksRepository.delete(idTasks);
-    }
-  }
-
   const boardsRepository = getRepository(Board);
-  const res = await boardsRepository.delete(id);
-  return res.raw;
+  const findIdBoard = await boardsRepository.find({ where: { id: id } });
+
+  if (findIdBoard[0]?.id !== undefined) {
+    const tasksRepository = getRepository(Task);
+    const arrayOfTasks = await tasksRepository.find({ where: { boardId: id } });
+
+    if (arrayOfTasks.length > 0) {
+      for (let i = 0; i < arrayOfTasks.length; i += 1) {
+        const idTasks = arrayOfTasks[i]!.id;
+        tasksRepository.delete(idTasks);
+      }
+    }
+
+    const res = await boardsRepository.delete(id);
+    return res.raw;
+  } else {
+    return undefined;
+  }
 };
 
 export const boardRepo = {
